+++ test/index.mjs
@@ -1,5 +1,6 @@
 import TestDirector from 'test-director';
 import testFindUnusedExportsCli from './cli/find-unused-exports.test.mjs';
+import testCliError from './private/CliError.test.mjs';
 import testGetVariableDeclarationIdentifierNames from './private/getVariableDeclarationIdentifierNames.test.mjs';
 import testIsDirectoryPath from './private/isDirectoryPath.test.mjs';
 import testReportCliError from './private/reportCliError.test.mjs';
@@ -11,6 +12,7 @@ const tests = new TestDirector();
 
 testReplaceStackTraces(tests);
 testFindUnusedExportsCli(tests);
+testCliError(tests);
 testGetVariableDeclarationIdentifierNames(tests);
 testIsDirectoryPath(tests);
 testReportCliError(tests);